export dts for window.xrCurrentSceneDefaults
