+++ .changeset/sad-dragons-check.md
@@ -1,0 +1,5 @@
+---
+'@webspatial/react-sdk': patch
+---
+
+export dts for window.xrCurrentSceneDefaults
